Rename test/spec to test/bench
