+++ test/bench.rb
@@ -1,6 +1,6 @@
 require_relative 'test_init'
 
 TestBench::Runner.(
-  'spec/**/*.rb',
+  'bench/**/*.rb',
   exclude_pattern: %r{/skip\.|(?:_init\.rb|\.sketch\.rb|_sketch\.rb|\.skip\.rb)\z}
 ) or exit 1
