+++ test/bench/bench_init.rb
