+++ test/bench/method_with_body.rb
@@ -1,4 +1,4 @@
-require_relative 'spec_init'
+require_relative './bench_init'
 
 module Fixture
   module MethodWithBody
