+++ test/bench/pure_virtual_method_is_defined.rb
@@ -1,4 +1,4 @@
-require_relative 'spec_init'
+require_relative './bench_init'
 
 module Fixture
   module PureVirtualMethodIsDefined
